Use numeric values instead of strings
Functions should return a numeric value, not a string.

The `font-size` property has been deleted from the `<html>` tag to avoid conflicts with other stylesheets. It will assume that the default `font-size` is `16px` (most web browsers set this value) and let the user change it value if they want.